Factor out LowLevelKeyFetcher
